+++ cloudnetpy/instruments/__init__.py
@@ -1,7 +1,7 @@
 from .basta import basta2nc
 from .ceilo import ceilo2nc
 from .copernicus import copernicus2nc
-from .disdrometer import disdrometer2nc
+from .disdrometer import parsivel2nc, thies2nc
 from .galileo import galileo2nc
 from .hatpro import hatpro2nc
 from .mira import mira2nc
