+++ cloudnetpy/instruments/disdrometer/__init__.py
@@ -1,0 +1,2 @@
+from .parsivel import parsivel2nc
+from .thies import thies2nc
