+++ cloudnetpy/instruments/disdrometer/common.py
@@ -4,70 +4,15 @@ import logging
 import numpy as np
 from numpy import ma
 
-from cloudnetpy import output, utils
+from cloudnetpy import utils
 from cloudnetpy.cloudnetarray import CloudnetArray
-from cloudnetpy.exceptions import DisdrometerDataError, ValidTimeStampError
-from cloudnetpy.instruments import instruments
+from cloudnetpy.exceptions import ValidTimeStampError
 from cloudnetpy.instruments.cloudnet_instrument import CloudnetInstrument
 from cloudnetpy.instruments.vaisala import values_to_dict
 from cloudnetpy.metadata import MetaData
 
 PARSIVEL = "OTT Parsivel-2"
 THIES = "Thies-LNM"
-
-
-def disdrometer2nc(
-    disdrometer_file: str,
-    output_file: str,
-    site_meta: dict,
-    uuid: str | None = None,
-    date: str | None = None,
-) -> str:
-    """Converts disdrometer data into Cloudnet Level 1b netCDF file. Accepts
-    measurements from OTT Parsivel-2 and Thies-LNM disdrometers.
-
-    Args:
-        disdrometer_file: Filename of disdrometer .log file.
-        output_file: Output filename.
-        site_meta: Dictionary containing information about the site. Required key
-            is `name`.
-        uuid: Set specific UUID for the file.
-        date: Expected date of the measurements as YYYY-MM-DD.
-
-    Returns:
-        UUID of the generated file.
-
-    Raises:
-        DisdrometerDataError: Timestamps do not match the expected date, or unable
-            to read the disdrometer file.
-
-    Examples:
-        >>> from cloudnetpy.instruments import disdrometer2nc
-        >>> site_meta = {'name': 'Lindenberg', 'altitude': 104, 'latitude': 52.2,
-        'longitude': 14.1}
-        >>> uuid = disdrometer2nc('thies-lnm.log', 'thies-lnm.nc', site_meta)
-
-    """
-    disdrometer: Thies | Parsivel
-    try:
-        disdrometer = Parsivel(disdrometer_file, site_meta)
-    except ValueError:
-        try:
-            disdrometer = Thies(disdrometer_file, site_meta)
-        except (ValueError, IndexError) as err:
-            raise DisdrometerDataError("Can not read disdrometer file") from err
-    if date is not None:
-        disdrometer.validate_date(date)
-    disdrometer.init_data()
-    if date is not None:
-        disdrometer.sort_timestamps()
-        disdrometer.remove_duplicate_timestamps()
-    disdrometer.add_meta()
-    disdrometer.convert_units()
-    attributes = output.add_time_attribute(ATTRIBUTES, disdrometer.date)
-    output.update_attributes(disdrometer.data, attributes)
-    uuid = output.save_level1b(disdrometer, output_file, uuid)
-    return uuid
 
 
 class Disdrometer(CloudnetInstrument):
@@ -253,160 +198,6 @@ class Disdrometer(CloudnetInstrument):
         bounds = np.stack((lower_limit, upper_limit)).T
         spread = bounds[:, 1] - bounds[:, 0]
         return mid_value, bounds, spread
-
-
-class Parsivel(Disdrometer):
-    def __init__(self, filename: str, site_meta: dict):
-        super().__init__(filename, site_meta, PARSIVEL)
-        self.n_velocity = 32
-        self.n_diameter = 32
-        self.date = self._init_date()
-        self._create_velocity_vectors()
-        self._create_diameter_vectors()
-        self.instrument = instruments.PARSIVEL2
-
-    def init_data(self):
-        """
-        Note:
-            This is a custom format submitted by Juelich, Norunda and Ny-Alesund
-            to Cloudnet data portal. It does not follow the order in the Parsivel2
-            manual https://www.fondriest.com/pdf/ott_parsivel2_manual.pdf
-        """
-        column_and_key = [
-            (0, "_time"),
-            (1, "rainfall_rate"),
-            (2, "_rain_accum"),
-            (3, "synop_WaWa"),
-            (4, "radar_reflectivity"),
-            (5, "visibility"),
-            (6, "interval"),
-            (7, "sig_laser"),
-            (8, "n_particles"),
-            (9, "T_sensor"),
-            (10, "_sensor_id"),  # to global attributes
-            (12, "I_heating"),
-            (13, "V_power_supply"),
-            (14, "state_sensor"),
-            (15, "_station_name"),
-            (16, "_rain_amount_absolute"),
-            (17, "error_code"),
-        ]
-        self._append_data(column_and_key)
-        self._append_vector_data()
-        self._append_spectra()
-
-    def _append_vector_data(self):
-        keys = ("number_concentration", "fall_velocity")
-        data = {
-            key: ma.masked_all((len(self._file_data["vectors"]), self.n_diameter))
-            for key in keys
-        }
-        for time_ind, row in enumerate(self._file_data["vectors"]):
-            values = _parse_int(row)
-            for key, array in zip(keys, np.split(values, 2)):
-                data[key][time_ind, :] = array
-        for key in keys:
-            self.data[key] = CloudnetArray(
-                data[key], key, dimensions=("time", "diameter")
-            )
-
-    def _init_date(self) -> list:
-        timestamp = self._file_data["scalars"][0][0]
-        return _parse_parsivel_timestamp(timestamp)[:3]
-
-    def _create_velocity_vectors(self):
-        n_values = [10, 5, 5, 5, 5, 2]
-        spreads = [0.1, 0.2, 0.4, 0.8, 1.6, 3.2]
-        self._store_vectors(n_values, spreads, "velocity")
-
-    def _create_diameter_vectors(self):
-        n_values = [10, 5, 5, 5, 5, 2]
-        spreads = [0.125, 0.25, 0.5, 1, 2, 3]
-        self._store_vectors(n_values, spreads, "diameter")
-
-
-class Thies(Disdrometer):
-    def __init__(self, filename: str, site_meta: dict):
-        super().__init__(filename, site_meta, THIES)
-        self.n_velocity = 20
-        self.n_diameter = 22
-        self.date = self._init_date()
-        self._create_velocity_vectors()
-        self._create_diameter_vectors()
-        self.instrument = instruments.THIES
-
-    def init_data(self):
-        """According to
-        https://www.biral.com/wp-content/uploads/2015/01/5.4110.xx_.xxx_.pdf
-        """
-        column_and_key = [
-            (1, "_serial_number"),
-            (2, "_software_version"),
-            (3, "_date"),
-            (4, "_time"),
-            (5, "_synop_5min_ww"),
-            (6, "_synop_5min_WaWa"),
-            (7, "_metar_5min_4678"),
-            (8, "_rainfall_rate_5min"),
-            (9, "synop_WW"),  # 1min
-            (10, "synop_WaWa"),  # 1min
-            (11, "_metar_1_min_4678"),
-            (12, "rainfall_rate_1min_total"),
-            (13, "rainfall_rate"),  # liquid, mm h-1
-            (14, "rainfall_rate_1min_solid"),
-            (15, "_precipition_amount"),  # mm
-            (16, "visibility"),
-            (17, "radar_reflectivity"),
-            (18, "measurement_quality"),
-            (19, "maximum_hail_diameter"),
-            (20, "status_laser"),
-            (21, "static_signal"),
-            (22, "status_T_laser_analogue"),
-            (23, "status_T_laser_digital"),
-            (24, "status_I_laser_analogue"),
-            (25, "status_I_laser_digital"),
-            (26, "status_sensor_supply"),
-            (27, "status_laser_heating"),
-            (28, "status_receiver_heating"),
-            (29, "status_temperature_sensor"),
-            (30, "status_heating_supply"),
-            (31, "status_heating_housing"),
-            (32, "status_heating_heads"),
-            (33, "status_heating_carriers"),
-            (34, "status_laser_power"),
-            (35, "_status_reserve"),
-            (36, "T_interior"),
-            (37, "T_laser_driver"),  # 0-80 C
-            (38, "I_mean_laser"),
-            (39, "V_control"),  # mV 4005-4015
-            (40, "V_optical_output"),  # mV 2300-6500
-            (41, "V_sensor_supply"),  # 1/10V
-            (42, "I_heating_laser_head"),  # mA
-            (43, "I_heating_receiver_head"),  # mA
-            (44, "T_ambient"),  # C
-            (45, "_V_heating_supply"),
-            (46, "_I_housing"),
-            (47, "_I_heating_heads"),
-            (48, "_I_heating_carriers"),
-            (49, "n_particles"),
-        ]
-        self._append_data(column_and_key)
-        self._append_spectra()
-
-    def _init_date(self) -> list:
-        first_date = self._file_data["scalars"][0][3]
-        first_date = _format_thies_date(first_date)
-        return first_date.split("-")
-
-    def _create_velocity_vectors(self):
-        n_values = [5, 6, 7, 1, 1]
-        spreads = [0.2, 0.4, 0.8, 1, 10]
-        self._store_vectors(n_values, spreads, "velocity")
-
-    def _create_diameter_vectors(self):
-        n_values = [3, 6, 13]
-        spreads = [0.125, 0.25, 0.5]
-        self._store_vectors(n_values, spreads, "diameter", start=0.125)
 
 
 def _parse_parsivel_timestamp(timestamp: str) -> list:
